Cache Matrix Inverse
Matrix inversion is usually a costly computation and there may be some benefit to caching the inverse of a matrix rather than compute it repeatedly (there are also alternatives to matrix inversion that we will not discuss here). The assignment here is to write a pair of functions that cache the inverse of a matrix.